dataIngestion/requirements.txt
couchbase
pandas
kafka-python